Move and improve matching functionality
